chore(future_builder): profile refactoring
