+++ lib/pages/profile/profile.dart
@@ -1,6 +1,7 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
+import 'package:remote_guide_firebase/pages/profile/profile_info.dart';
 
 class Profile extends StatefulWidget {
   const Profile({Key? key}) : super(key: key);
@@ -26,7 +27,7 @@ class _ProfileState extends State<Profile> {
         style: Theme.of(context).textTheme.displayMedium!,
         textAlign: TextAlign.center,
         child:
-        FutureBuilder<Object?>(
+          FutureBuilder<Object?>(
             future: _data,
             builder: (BuildContext context, AsyncSnapshot<Object?> snapshot) {
               if (snapshot.connectionState == ConnectionState.done) {
@@ -40,54 +41,6 @@ class _ProfileState extends State<Profile> {
     );
   }
 }
-
-class ProfileInfo extends StatefulWidget {
-  final data;
-  const ProfileInfo(this.data, {super.key});
-
-  @override
-  State<ProfileInfo> createState() => _ProfileInfoState();
-}
-
-class _ProfileInfoState extends State<ProfileInfo> {
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Column(
-        children: [
-          Padding(
-            padding: const EdgeInsets.only(left: 20.0, top: 8.0),
-            child: Row(
-              children: [
-                Column(
-                  mainAxisAlignment: MainAxisAlignment.start,
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: const [
-                    Text("Name"),
-                    Text("Email"),
-                    Text("Status"),
-                  ],
-                ),
-                Padding(
-                    padding: const EdgeInsets.fromLTRB(20.0, 0.0, 0.0, 0.0),
-                    child: Column(
-                      children: [
-                        Text(widget.data.get('name')),
-                        Text(widget.data.get('email')),
-                        Text(widget.data.get('guide') ? "Guide" : "Client"),
-                      ],
-                    )
-                )
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
 
 //   @override
 //   Widget build(BuildContext context) {
